+++ ConsumerApi.Sdk/DeviceData.cs
@@ -1,0 +1,9 @@
+using Backbone.ConsumerApi.Sdk.Authentication;
+
+namespace Backbone.ConsumerApi.Sdk;
+
+public record DeviceData
+{
+    public required string DeviceId;
+    public required UserCredentials UserCredentials;
+}
