+++ ConsumerApi.Sdk/IdentityData.cs
@@ -1,0 +1,9 @@
+using Backbone.Crypto;
+
+namespace Backbone.ConsumerApi.Sdk;
+
+public record IdentityData
+{
+    public required string Address;
+    public required KeyPair KeyPair;
+}
